+++ cgc/coclustering.py
@@ -15,19 +15,59 @@ logger = logging.getLogger(__name__)
 
 class CoclusteringResults(Results):
     """
-    Contains results and metadata of a co-clustering calculation
+    Contains results and metadata of a co-clustering calculation.
+
+    :var row_clusters: Final row cluster assignment.
+    :type row_clusters: numpy.ndarray
+    :var col_clusters: Final column cluster assignment.
+    :type col_clusters: numpy.ndarray
+    :var error: Approximation error of the co-clustering.
+    :type error: float
+    :var nruns_completed: Number of successfully completed runs.
+    :type nruns_completed: int
+    :var nruns_converged: Number of converged runs.
+    :type nruns_converged: int
     """
-    def reset(self):
-        self.row_clusters = None
-        self.col_clusters = None
-        self.error = None
-        self.nruns_completed = 0
-        self.nruns_converged = 0
+    row_clusters = None
+    col_clusters = None
+    error = None
+    nruns_completed = 0
+    nruns_converged = 0
 
 
 class Coclustering(object):
     """
-    Perform the co-clustering analysis of a 2D array
+    Perform a co-clustering analysis for a two-dimensional array.
+
+    :param Z: Data matrix for which to run the co-clustering analysis
+    :type Z: numpy.ndarray or dask.array.Array
+    :param nclusters_row: Number of row clusters.
+    :type nclusters_row: int
+    :param nclusters_col: Number of column clusters.
+    :type nclusters_col: int
+    :param conv_threshold: Convergence threshold for the objective function.
+    :type conv_threshold: float, optional
+    :param max_iterations: Maximum number of iterations.
+    :type max_iterations: int, optional
+    :param nruns: Number of differently-initialized runs.
+    :type nruns: int, optional
+    :param output_filename: Name of the JSON file where to write the results.
+    :type output_filename: string, optional
+    :param row_clusters_init: Initial row cluster assignment.
+    :type row_clusters_init: numpy.ndarray or array_like, optional
+    :param col_clusters_init: Initial column cluster assignment.
+    :type col_clusters_init: numpy.ndarray or array_like, optional
+
+    :Example:
+
+    >>> import numpy as np
+    >>> Z = np.random.randint(1, 100, size=(10, 8)).astype('float64')
+    >>> cc = Coclustering(Z,
+                          nclusters_row=5,
+                          nclusters_col=4,
+                          max_iterations=50,
+                          nruns=10)
+
     """
     def __init__(self,
                  Z,
@@ -36,31 +76,15 @@ class Coclustering(object):
                  conv_threshold=1.e-5,
                  max_iterations=1,
                  nruns=1,
-                 epsilon=1.e-8,
                  output_filename='',
                  row_clusters_init=None,
                  col_clusters_init=None):
-        """
-        Initialize the object
-
-        :param Z: m x n data matrix
-        :param nclusters_row: number of row clusters
-        :param nclusters_col: number of column clusters
-        :param conv_threshold: convergence threshold for the objective function
-        :param max_iterations: maximum number of iterations
-        :param nruns: number of differently-initialized runs
-        :param epsilon: numerical parameter, avoids zero arguments in log
-        :param output_filename: name of the file where to write the clusters
-        :param row_clusters_init: initial row clusters
-        :param col_clusters_init: initial column clusters
-        """
         # Input parameters -----------------
         self.nclusters_row = nclusters_row
         self.nclusters_col = nclusters_col
         self.conv_threshold = conv_threshold
         self.max_iterations = max_iterations
         self.nruns = nruns
-        self.epsilon = epsilon
         self.output_filename = output_filename
         self.row_clusters_init = row_clusters_init
         self.col_clusters_init = col_clusters_init
@@ -80,11 +104,16 @@ class Coclustering(object):
 
     def run_with_dask(self, client=None, low_memory=True):
         """
-        Run the co-clustering with Dask
+        Run the co-clustering analysis using Dask.
 
-        :param client: Dask client
-        :param low_memory: if false, all runs are submitted to the Dask cluster
-        :return: co-clustering results
+        :param client: Dask client. If not specified, the default
+            `LocalCluster` is employed.
+        :type client: dask.distributed.Client, optional
+        :param low_memory: If False, all runs are submitted to the Dask cluster
+            (experimental feature, discouraged).
+        :type low_memory: bool, optional
+        :return: Co-clustering results.
+        :type: cgc.coclustering.CoclusteringResults
         """
         self.client = client if client is not None else Client()
 
@@ -96,19 +125,19 @@ class Coclustering(object):
         self.results.write(filename=self.output_filename)
         return self.results
 
-    def run_with_threads(self,
-                         nthreads=1,
-                         low_memory=False,
-                         numba_jit=False):
+    def run_with_threads(self, nthreads=1, low_memory=False):
         """
-        Run the co-clustering using an algorithm based on numpy + threading
-        (only suitable for local runs)
+        Run the co-clustering using an algorithm based on Numpy plus threading
+        (only suitable for local runs).
 
-        :param nthreads: number of threads
-        :param low_memory: if true, use a memory-conservative algorithm
-        :param numba_jit: if true, and low_memory is true, then use Numba
-                          just-in-time compilation to improve performance
-        :return: co-clustering results
+        :param nthreads: Number of threads employed to simultaneously run
+            differently-initialized co-clustering analysis.
+        :type nthreads: int, optional
+        :param low_memory: Make use of a low-memory version of the algorithm
+            with Numba JIT acceleration
+        :type low_memory: bool, optional
+        :return: Co-clustering results.
+        :type: cgc.coclustering.CoclusteringResults
         """
         with ThreadPoolExecutor(max_workers=nthreads) as executor:
             futures = [
@@ -118,9 +147,7 @@ class Coclustering(object):
                                 self.nclusters_col,
                                 self.conv_threshold,
                                 self.max_iterations,
-                                self.epsilon,
                                 low_memory,
-                                numba_jit,
                                 row_clusters_init=self.row_clusters_init,
                                 col_clusters_init=self.col_clusters_init)
                 for _ in range(self.nruns)
@@ -143,7 +170,7 @@ class Coclustering(object):
         return self.results
 
     def _dask_runs_memory(self):
-        """ Memory efficient Dask implementation: sequential runs """
+        """ Memory efficient Dask implementation: sequential runs. """
         for r in range(self.nruns):
             logger.info(f'Run {self.results.nruns_completed}')
             converged, niters, row, col, e = coclustering_dask.coclustering(
@@ -152,10 +179,8 @@ class Coclustering(object):
                 self.nclusters_col,
                 self.conv_threshold,
                 self.max_iterations,
-                self.epsilon,
                 row_clusters_init=self.row_clusters_init,
-                col_clusters_init=self.col_clusters_init
-            )
+                col_clusters_init=self.col_clusters_init)
             logger.info(f'Error = {e}')
             if converged:
                 logger.info(f'Run converged in {niters} iterations')
@@ -171,22 +196,21 @@ class Coclustering(object):
     def _dask_runs_performance(self):
         """
         Faster but memory-intensive Dask implementation: all runs are
-        simultaneosly submitted to the scheduler
+        simultaneously submitted to the scheduler (experimental, discouraged).
         """
         Z = self.client.scatter(self.Z)
-        futures = [self.client.submit(
-                       coclustering_dask.coclustering,
-                       Z,
-                       self.nclusters_row,
-                       self.nclusters_col,
-                       self.conv_threshold,
-                       self.max_iterations,
-                       self.epsilon,
-                       row_clusters_init=self.row_clusters_init,
-                       col_clusters_init=self.col_clusters_init,
-                       run_on_worker=True,
-                       pure=False)
-                   for _ in range(self.nruns)]
+        futures = [
+            self.client.submit(coclustering_dask.coclustering,
+                               Z,
+                               self.nclusters_row,
+                               self.nclusters_col,
+                               self.conv_threshold,
+                               self.max_iterations,
+                               row_clusters_init=self.row_clusters_init,
+                               col_clusters_init=self.col_clusters_init,
+                               run_on_worker=True,
+                               pure=False) for _ in range(self.nruns)
+        ]
         for future, result in dask.distributed.as_completed(futures,
                                                             with_results=True):
             logger.info(f'Retrieving run {self.results.nruns_completed} ..')